reorganized and renamed functions and files based on library
